requirements.txt
autopep8
pylint
argparse
ffmpeg-python